+++ js/card.js
@@ -1,6 +1,13 @@
 const cardsList = document.querySelector('.map__canvas');
 const cardTemplate = document.querySelector('#card').content.querySelector('.popup');
 
+const TYPES_OF_HOUSING = {
+  palace: 'Дворец',
+  flat: 'Квартира',
+  house: 'Дом',
+  bungalow: 'Бунгало',
+  hotel: 'Отель',
+};
 
 function createCard(object) {
   const cardElement = cardTemplate.cloneNode(true);
@@ -8,7 +15,7 @@ function createCard(object) {
   cardElement.querySelector('.popup__title').textContent = object.offer.title;
   cardElement.querySelector('.popup__text--address').textContent = object.offer.address;
   cardElement.querySelector('.popup__text--price').textContent = `${object.offer.price} ₽/ночь`;
-  cardElement.querySelector('.popup__type').textContent = object.offer.type.display;
+  cardElement.querySelector('.popup__type').textContent = TYPES_OF_HOUSING[object.offer.type.display];
   cardElement.querySelector('.popup__text--capacity').textContent = `${object.offer.rooms} комнаты для ${object.offer.guests} гостей`;
   cardElement.querySelector('.popup__text--time').textContent = `Заезд после ${object.offer.checkin}, выезд до ${object.offer.checkout}`;
   cardElement.querySelector('.popup__description').textContent = object.offer.description;
@@ -21,13 +28,15 @@ function createCard(object) {
   }
   //Создает список удобств
   const modifiers = object.offer.features.map((feature) => `popup__feature--${feature}`);
-  object.innerHTML = '';
   cardElement.querySelectorAll('.popup__feature').forEach((item) => {
     const modifier = item.classList[1];
     if (!modifiers.includes(modifier)) {
       item.remove();
     }
   });
+
+  const photos = cardElement.querySelector('.popup__photos');
+  photos.innerHTML = '';
 
   // Добавляет фото
   const photoCardElement = document.createElement('img');
@@ -43,4 +52,8 @@ function createCard(object) {
   return cardElement;
 }
 
-export {createCard, cardsList};
+const renderCard = (card) => {
+  cardsList.appendChild(card);
+};
+
+export {createCard, renderCard};
