+++ js/main.js
@@ -1,15 +1,8 @@
-import {similarAd} from './data.js';
-import {createCard, cardsList} from './card.js';
+import {createAds} from './data.js';
+import {createCard, renderCard} from './card.js';
 
 
-const SIMILAR_COUNT = 10;
-similarAd(SIMILAR_COUNT);
-
-const createAds = () => new Array(SIMILAR_COUNT).fill(null).map(() => createAd());
-
-function createCards(arr, container) {
-  arr.forEach((item) => container.append(createCard(item)));
-}
-
-createCards(createAds(), cardsList);
-
+const COUNT = 10;
+const array = createAds(COUNT);
+const card = createCard(array[0]);
+renderCard(card);
